Ajouter le tri par tas itératif

+++ src/miniproject2/sort/BubbleSortAlgorithm.java
@@ -11,7 +11,7 @@ public class BubbleSortAlgorithm extends SortAlgorithm {
 
     @Override
     protected void doOperation() {
-        long progress = 0;
+        int progress = 0;
         for(int i = 0; i < numbers.length - 1; i++){
             for (int j = i + 1; j < numbers.length; j++){
                 if(numbers[i] > numbers [j]){
